% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char.R
\name{char}
\alias{char}
\alias{char,character-method}
\alias{char,array-method}
\title{Convert an array to a character array}
\usage{
char(A)

\S4method{char}{character}(A)

\S4method{char}{array}(A)
}
\arguments{
\item{A}{a vector or array (not yet supported)}
}
\value{
\code{A} converted to characters
}
\description{
A character array is a sequence of characters, just as a
numeric array is a sequence of numbers. A typical use is to store a short
piece of text as a row of characters in a character vector.
}
\section{Methods (by class)}{
\itemize{
\item \code{char(character)}: Converting a character vector

\item \code{char(array)}: Converting a character array

}}
\examples{
char("Hi!")
char(matrix(letters, 2))
}
\author{
Waldir Leoncio
}
