% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colon.R
\name{colon}
\alias{colon}
\title{Vector creation}
\usage{
colon(a, b)
}
\arguments{
\item{a}{initial number}

\item{b}{final number}
}
\value{
A vector containing a sequence of integers going from a to b
}
\description{
Simulates the function \code{colon()} and its equivalent \code{:}
operator from Matlab, which have a similar but not quite equivalent behavior
when compared to \code{seq()} and \code{:} in R.
}
\examples{
colon(1, 4)
colon(4, 8)
}
