% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{find}
\alias{find}
\title{Find indices and values of nonzero elements}
\usage{
find(x, sort = TRUE)
}
\arguments{
\item{x}{object or logic operation on an object}

\item{sort}{sort output?}
}
\value{
A vector of indices of x that satisfy the logical test
(nonzero, by default).
}
\description{
Emulates behavior of \code{find}
}
\examples{
X <- matrix(c(1, 0, 2, 0, 1, 1, 0, 0, 4), 3, byrow = TRUE)
Y <- seq(1, 19, 2)
find(X)
find(Y == 13)
}
