% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.R
\name{fix}
\alias{fix}
\title{Round toward zero}
\usage{
fix(X)
}
\arguments{
\item{X}{input element}
}
\value{
The values of \code{trunc(X)}.
}
\description{
Rounds each element of input to the nearest integer towards
zero. Basically the same as trunc()
}
\examples{
X <- matrix(c(-1.9, -3.4, 1.6, 2.5, -4.5, 4.5), 3, byrow = TRUE)
Y <- matrix(c(-1, -3, 1, 2, -4, 4), 3, byrow = TRUE)
fix(X)
fix(Y)
}
\author{
Waldir Leoncio
}
