% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammaln.R
\name{gammaln}
\alias{gammaln}
\title{Logarithm of gamma function}
\usage{
gammaln(A)
}
\arguments{
\item{A}{a non-negative, real matrix, vector or scalar}
}
\value{
An element-by-element \code{ln(gamma())}-transformed A
}
\description{
Calculates the natural logarithm of the gamma function
}
\note{
For MATLAB output reproduction, non-positive values will be
}
\examples{
gammaln(8)
gammaln(0)
gammaln(matrix(1:9, 3))
gammaln(-4:10)
}
\author{
Waldir Leoncio
}
