% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isFilePath.R
\name{isFilePath}
\alias{isFilePath}
\title{Check if an input is a valid path}
\usage{
isFilePath(x)
}
\arguments{
\item{x}{input}
}
\value{
\code{TRUE} if \code{x} is a valid path, \code{FALSE} otherwise
}
\description{
A simple check if an input corresponds to a valid path to a file
}
\author{
Waldir Leoncio
}
