% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linspace.R
\name{linspace}
\alias{linspace}
\title{Generate linearly-spaced vector}
\usage{
linspace(x1, x2, n = 100L)
}
\arguments{
\item{x1}{start point}

\item{x2}{end point}

\item{n}{length of output}
}
\value{
A numeric vector of \code{n} numbers between \code{x1} and \code{x2}.
}
\description{
This is a soft wrap around the \code{\link[base:seq]{base::seq()}} function.
}
\examples{
linspace(-5, 4)
linspace(1 + 2i, 9 + 9i, 5)
}
\author{
Waldir Leoncio
}
