% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log2.R
\name{log2}
\alias{log2}
\title{Base 2 logarithm}
\usage{
log2(X, tol = 1e-04)
}
\arguments{
\item{X}{a scalar or vector of numbers}
}
\value{
a list of mantissas and exponents such that mantissa * 2 ^ exponent
equals X
}
\description{
Base 2 logarithm and floating-point number dissection
}
\note{
User, beware: this is ridiculously slow and written in a hurry. If you
simply want to compute the base 2 logarithm of a number, use \code{log2()}
from \code{base}.
}
