% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_max.R
\name{min}
\alias{min}
\title{Minimum (MATLAB version)}
\usage{
min(X, indices = TRUE)
}
\arguments{
\item{X}{matrix}

\item{indices}{return indices?}
}
\value{
Either a list or a vector
}
\description{
Finds the minimum value for each column of a matrix, potentially returning the indices instead
}
\examples{
A <- matrix(c(23, 42, 37, 15, 52))
min(A)
base::min(A) # for comparison
}
\author{
Waldir Leoncio
}
