% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rem.R
\name{rem}
\alias{rem}
\title{Remainder after division½}
\usage{
rem(a, b)
}
\arguments{
\item{a}{the dividend}

\item{b}{the divisor}
}
\value{
The remainder
}
\description{
Rreturns the remainder after division of \code{a} by \code{b},
where \code{a} is the dividend and \code{b} is the divisor. This function is
often called the remainder operation. The \code{rem} function follows the
convention that \code{rem(a,0)} is \code{NaN}.
}
\examples{
rem(23, 5)
rem(1:5, 3)
rem(c(-4, -1, 7, 9), 3)  #FIXME
rem(c(0, 3.5, 5.9, 6.2, 9, 4 * pi), 2 * pi)
}
\author{
Waldir Leoncio
}
