% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setdiff.R
\name{setdiff}
\alias{setdiff}
\title{Set differences of two arrays}
\usage{
setdiff(A, B, legacy = FALSE)
}
\arguments{
\item{A}{first array}

\item{B}{second array}

\item{legacy}{if \code{TRUE}, preserves the behavior of the setdiff function from
MATLAB R2012b and prior releases. (currently not supported)}
}
\value{
An array containing he elements which are in A but not in B
}
\description{
Loosely replicates the behavior of the homonym Matlab function
}
\examples{
A <- c(3, 6, 2, 1, 5, 1, 1)
B <- c(2, 4, 6)
setdiff(A, B)
}
\author{
Waldir Leoncio
}
