% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortrows.R
\name{sortrows}
\alias{sortrows}
\title{Sort rows of matrix or table}
\usage{
sortrows(A, column = 1)
}
\arguments{
\item{A}{matrix}

\item{column}{ordering column}
}
\value{
The A matrix sorted by the first row, then the second
}
\description{
Emulates the behavior of the \code{sortrows} function on Matlab
}
\examples{
mx <- matrix(c(3, 2, 2, 1, 1, 10, 0, pi), 4)
mx
sortrows(mx)
}
