% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disp.R
\name{disp}
\alias{disp}
\title{Display the value of a variable}
\usage{
disp(X)
}
\arguments{
\item{X}{variable}
}
\value{
The value of \code{X}
}
\description{
\code{disp(X)} displays the value of variable \code{X}
without printing the variable name. This is a wrapper around \code{base::cat()}
that includes a breakline in the end.
}
\examples{
A <- c(15, 150)
S <- 'Hello World.'
disp(A)
disp(S)
}
\author{
Waldir Leoncio
}
